Code review

+++ Utils/WString.cpp
@@ -1,9 +1,8 @@
 #define _CRT_SECURE_NO_WARNINGS
-
 #include "WString.h"
 
 wstring WString::GetSymbolFromCString(const char* str, USHORT index) {
-	byte messageLen = strlen(str);
+	byte messageLen = static_cast<byte>(strlen(str));
 	char* messageDup = new char[messageLen + 1];
 	strncpy(messageDup, str + index, (messageLen - index) + 1);
